[Add] DeleteThreadsTest - only_thread_creators_can_delete_his_threads

+++ app/Http/Controllers/ThreadController.php
@@ -47,6 +47,8 @@ class ThreadController extends Controller
 
     public function destroy(Thread $thread)
     {
+        $this->authorize('delete', $thread);
+
         $thread->delete();
 
         return redirect('/users/' . auth()->id());
